Fix pagination issue
In `DataFetchingContainer.js` the paging token is always taken
from `records[0]`. The problem is that the records in a prev
request descend and only this wrapper reverses the list.

So without this wrapper the paging token for prev requests is
always taken from the top most, most recent record. If one
then presses the prev button another time one actually moves
forward instead of further backward.