Refactor fetchRecords for OperationTable
This refactoring puts the builder on which `.call()` has not yet
been invoked into a dedicated variable. In the follow-up commit
to this one, we aim to repeatedely fetch pages, filter them, and
aggregate their result. Therefore it is necessary to increase
the cursor on the builder each time a new page is fetched.

There is a bug in js-stellar-sdk though:
The current behavior of js-stellar-sdk is that only the first
cursor which is set on the builder has an effect -- all latter
invocations of `.cursor()` won't have an effect.
This refactoring is necessary because after `.call()` there would
be a (default) cursor set on the builder.

This behavior of only the first call having an effect is fixed in
https://github.com/stellar/js-stellar-sdk/pull/193, but until the
fix is released this workaround is necessary. It gives the
possiblity to return an "unfinished" builder on which it is then
possible to call `.cursor(newCursor).call()`.