Add possibility to filter operations for operation type
Horizon does not support filtering `/operations` for an
operation type, hence this wrapper is used.
The wrapper basically works around the missing filtering
in Horizon and fetches data page by page, filtering it
each time until the configured limit of records per page
is reached.

Unfortunately this has limitations, because basically the
halting problem occurs: There is no way for us to know how
many records exist for a certain operation type and hence
we don't know when to stop fetching.

For example, somebody could create a filter for ops of type
`foo`. If the total operations dataset is very large, but
contains only one record of type `foo` then we would
recursively fetch the entire operations dataset, because we
don't know when to stop fetching records and because we want
to fill the page with records. So even after we fetched the
only one record we would still need to search the entire
dataset and look if there is another matching record.

To counter this problem there is a hard limit after which
fetching is stopped and a disclaimer is displayed.